requirements.txt
google-generativeai>=0.8.0
fastmcp>=0.9.0
streamlit>=1.28.0
python-dotenv>=1.0.0
ollama>=0.1.7

